+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.6.0
 + Implement `dropLast(n)`
-+ Add support for `zipWith(collection)` and `zipWith(iterator)`
++ Add support for `zipWith(iterable)` and `zipWith(iterator)`
++ Add support for `interleave(iterable)` and `interleave(iterator)`
 
 ### 0.5.0
 + Implement `reverse()`